requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.1
idna==3.7
psycopg2==2.9.9
PyJWT==2.8.0
pyTelegramBotAPI==4.17.0
python-environ==0.4.54
redis==5.0.6
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
